+++ MainSource/Util.h
@@ -11,6 +11,6 @@ class Util{
     
         Util(){}
         static void splitString(std::vector<std::string>& returned ,std::string originalString, std::string delimiter);
-        static bool ChooseTwoProductIndexes(int &first, int &second, int numProducts, const std::set<std::pair<int,int> > & swapsDone);
+        static bool ChooseTwoProductIndexes(int &first, int &second, unsigned long long numProducts, const std::set<std::pair<int,int> > & swapsDone);
 };
